+++ docs/.vitepress/theme/index.ts
@@ -3,6 +3,7 @@ import DefaultTheme from 'vitepress/theme'
 
 import DemoGkAlert from '../components/demos/alert/DemoGkAlert.vue'
 import DemoGkButton from '../components/demos/button/DemoGkButton.vue'
+import DemoGkBottomSheet from '../components/demos/containment/bottom-sheet/DemoGkBottomSheet.vue'
 import DemoGkDialog from '../components/demos/containment/dialog/DemoGkDialog.vue'
 import DemoGkOverlay from '../components/demos/containment/overlay/DemoGkOverlay.vue'
 import DemoGkCheckbox from '../components/demos/form/checkbox/DemoGkCheckbox.vue'
@@ -30,6 +31,7 @@ export default {
     app.component('DemoGkCheckbox', DemoGkCheckbox)
     app.component('DemoGkRadio', DemoGkRadio)
     app.component('DemoGkSelect', DemoGkSelect)
+    app.component('DemoGkBottomSheet', DemoGkBottomSheet)
     app.component('DemoGkDialog', DemoGkDialog)
     app.component('DemoGkOverlay', DemoGkOverlay)
     app.component('DemoGkAlert', DemoGkAlert)