+++ src/vue/layout.ts
@@ -1,6 +1,7 @@
 export { default as GkAlert } from './components/alert/GkAlert.vue'
 export { default as GkOverlay } from './components/containment/overlay/GkOverlay.vue'
 export { default as GkDialog } from './components/containment/dialog/GkDialog.vue'
+export { default as GkBottomSheet } from './components/containment/bottom-sheet/GkBottomSheet.vue'
 export { default as GkStack } from './components/stack/GkStack.vue'
 export { default as GkContainer } from './components/container/GkContainer.vue'
 export { default as GkDivider } from './components/divider/GkDivider.vue'